merge with upstream

+++ R/BuildVignettes.R
@@ -33,11 +33,14 @@ BuildVignettes <- function(dir=getwd(), doc=file.path(dir, "inst/doc"),
                            gs_quality=c("ebook", "printer", "screen", "none"),
                            clean=TRUE, quiet=TRUE) {
 
-  checkmate::assertFileExists(file.path(dir, "DESCRIPTION"))
+  checkmate::assertString(dir)
   checkmate::assertPathForOutput(doc, overwrite=TRUE)
   gs_quality <- match.arg(gs_quality)
   checkmate::assertFlag(clean)
   checkmate::assertFlag(quiet)
+
+  dir <- normalizePath(path.expand(dir), winslash="/", mustWork=FALSE)
+  checkmate::assertFileExists(file.path(dir, "DESCRIPTION"))
 
   tools::buildVignettes(dir=dir, quiet=quiet, clean=clean, tangle=TRUE)
 
@@ -48,7 +51,7 @@ BuildVignettes <- function(dir=getwd(), doc=file.path(dir, "inst/doc"),
   if (v$dir != doc) {
     dir.create(doc, showWarnings=!quiet, recursive=TRUE)
     file.copy(c(v$docs, out), doc, overwrite=TRUE)
-    file.remove(out)
+    if (clean) file.remove(out)
   }
 
   if (gs_quality != "none") {
